fix: make src matching work again, reintroduce custom url input

+++ entrypoint.sh
@@ -1,14 +1,20 @@
 #!/bin/bash
 
+if [[ $# -ne 1 ]]; then
+  echo "Exactly one parameter required: idea-url"
+  exit 1
+fi
+
 git config --global --add safe.directory /github/workspace
 cd /github/workspace
 
+IDEA_URL=$1
 IDEA_DIR=".idea-cache"
 mkdir -p "$IDEA_DIR"
 
 if [[ ! -d "$IDEA_DIR/bin" ]]; then
   echo "Download IntelliJ IDEA..."
-  wget --no-verbose -O /tmp/idea.tar.gz "https://download.jetbrains.com/idea/ideaIC-2025.2.6.tar.gz"
+  wget --no-verbose -O /tmp/idea.tar.gz "$IDEA_URL"
   tar xzf /tmp/idea.tar.gz -C "$IDEA_DIR" --strip-components=1
   rm /tmp/idea.tar.gz
 fi
@@ -17,7 +23,13 @@ version=$("$IDEA_DIR/bin/idea.sh" --version 2>/dev/null || true)
 echo "Running:"
 echo "$version"
 echo "Formatting..."
-"$IDEA_DIR/bin/format.sh" -m "./src/**/*.java" -s "./style/IntelliJ IDEA.xml" -r .
 
-echo "Finished"
+files=$(find ./src -type f -name "*.java")
+"$IDEA_DIR/bin/format.sh" -m "$files" -s "./style/IntelliJ IDEA.xml" -r .
+
+echo "::notice title=Finished"
 git --no-pager diff --exit-code
+if [[ "$?" != "0" ]]; then
+  echo "::warning title=Style Checks failed"
+  exit 1
+fi
